// src/lib/nest-new/nest-new.schema.d.ts
export interface NestNewOptions {
    /**
     * Nest application name.
     */
    // NOTE: We could have a `number` here due to this issue: https://github.com/nestjs/nest-cli/issues/1519
    name: string | number;
    /**
     * Nest application author.
     */
    author?: string;
    /**
     * Nest application description.
     */
    description?: string;
    /**
     * Nest application destination directory.
     */
    directory?: string;
    /**
     * With TypeScript strict mode.
     */
    strict?: boolean;
    /**
     * Nest application version.
     */
    version?: string;
    /**
     * Application language.
     */
    language?: string;
    /**
     * The used package manager.
     */
    packageManager?: 'npm' | 'yarn' | 'pnpm' | 'undefined';
    /**
     * Nest included production dependencies (comma separated values).
     */
    dependencies?: string;
    /**
     * Nest included development dependencies (comma separated values).
     */
    devDependencies?: string;
    /**
     * Create a new initial application in the 'src' folder of the new workspace.
     */
    createApplication?: boolean;
}